Most of the tart_spi has been integrated into tart_1.0 IP.
The SPI bus has been replaced with an AXI bus which is not
yet fully implimented. Currently it works as a delayed response
to 4 registers. Next it will need to talk over the wishbone
bus to read and write register values.

SRAM or some kind of replacement has not yet been implimented.

Clocks are genereted by Xilinx Clock Wizard with connects to the
tart IP as I couldn't figure out how to put that in the tart IP.