AXI DMA implimented.
There seems to be a bug with enabling the correlator that causes
pl to hang.

DMA times out the first time the module is loaded. I am not sure
why as pl gets reset by the module.

The files zynq uses that are different from tart_spi will be moved
to the zynq directory and this branch will be merged into master
soon.

PL side of zynq support is almost done.