tart_hardware_interface does basic zynq functions.
Register reading and writing works.
Reading acquisition data should be working but I have
not tested it.

Currently tart_testbench gets hung up after
start_acquisition. It seems that the tart never comes
out of pause.

How should detecting if the software is running on a
zynq or a raspberry pi work? As is you will need to
modify tarthw.py in tart_hardware_interface to create
a TartAXI class for the zynq or a TartSPI class for the
raspberry pi.