tart_hardware_interface works as I expect it to.
Which means it doesn't work and I don't understand
how it should work.

When run with --internal and --counter it gets an
a stream of data that increments every sixth
word which makes sense as fake_telescope on the
buffered clock and the AXI bus runs on the 6x
clock. How this worked in the raspberry pi I am
not sure.